Add ability to get process message
